+++ akshare/__init__.py
@@ -2404,9 +2404,10 @@ amac_manager_cancelled_info # 中国证券投资基金业协会-信息公示-诚
 1.9.86 fix: fix index_kq_fz interface
 1.9.87 add: add option_minute_em interface
 1.9.88 fix: fix setup.py
+1.9.89 fix: fix index_kq_fz interface
 """
 
-__version__ = "1.9.88"
+__version__ = "1.9.89"
 __author__ = "AKFamily"
 
 import sys