+++ docs/changelog.md
@@ -56,6 +56,10 @@
 
 ## 更新说明
 
+1.9.89 fix: fix index_kq_fz interface
+
+    1. 修复 index_kq_fz 接口
+
 1.9.88 fix: fix setup.py
     
     1. 移除 setup.py 中的相关依赖
@@ -2333,6 +2337,8 @@
     3. 修改 air_quality_hist 接口的请求日期格式，从 '2022-01-01' 为 '20220101'
 
 ## 版本更新说明
+
+1.9.89 fix: fix index_kq_fz interface
 
 1.9.88 fix: fix setup.py